Adjust Layout on Homepage
